enhancement: Re-enable legacy edit tiers page
