+++ components/admin-panel/constants.js
@@ -64,6 +64,7 @@ export const COLLECTIVE_SECTIONS = {
   TEAM: 'members',
   TICKETS: 'tickets',
   TIERS: 'tiers',
+  TIERS_LEGACY: 'tiers-legacy',
   TWO_FACTOR_AUTH: 'two-factor-auth',
   VIRTUAL_CARDS: 'virtual-cards',
   WEBHOOKS: 'webhooks',
