+++ components/edit-collective/Menu.js
@@ -29,6 +29,7 @@ export const EDIT_COLLECTIVE_SECTIONS = {
   PAYMENT_RECEIPTS: 'payment-receipts',
   TICKETS: 'tickets',
   TIERS: 'tiers',
+  TIERS_LEGACY: 'tiers-legacy',
   GIFT_CARDS: 'gift-cards',
   AUTHORIZED_APPS: 'authorized-apps',
   FOR_DEVELOPERS: 'for-developers',
