+++ components/edit-collective/sections/Tickets.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Tiers from './Tiers';
+import Tiers from './TiersLegacy';
 
 const Tickets = props => <Tiers {...props} types={['TICKET']} defaultType="TICKET" />;
